add tables and models
restaurants >has_many|belongs_to< tables
tables >has_many|belongs_to< reservations
tables >has_many|belongs_to< availabilities

==============================

initially went for tables has_many reservations through availabilities
but decided to separate relations. It is convenient to have table_id foreign_key in
reservations for the query simplicity. Also it is convenient to have full reservation
info in the table itself when querying. Later we might have multiple constraints similar
to availability and having through relation for each of them could be a pain.
Downside is that we have :date both in availabilities and reservations tables
which is slight denormalization.

==============================

I don't usually spec default rails validation like simple presence and uniqueness, associations etc
but I do usually spec all custom validations and/or conditions. omitted in this task.
It's preferable to avoid model callbacks to avoid callback hell but model callbacks are used in
this task for simplicity.

==============================

table name tables is terrible, but I went with it anyway for this task

==============================

added factories as well with minimum valid amount of args