add availability query
I usually transcribe generated SQL for complex queries like I did here.
Alternatively the query could have been suitable_tables left outer join availabilities_to_date,
which would probably remove one select from select. Had to select tables.id as _table_id to prevent
rails overriding the availabilities :id in the resulting objects which would lead to the availability
instances having :table_id in the :id field.

==============================

The query includes 4 range filters, this is tough to index. Only one index for the range filters will
be in use. Added 2 column compound index to tables to optimize query. It's restaurant_id and one of
the columns with a range filter on it. Ideally it should be the one that filters out the most records.
Explain shows usage of the index and lesser complexity.

==============================

Regarding the specs: some devs just check the resulting sql and call it a day:
expect(query.to_sql).to include('WHERE something IS NULL')
I also do this for the simple active record only queries which basically just conditionally apply filters.

Here this is a vital query for our app, and it should be tested fully.
This is a heavy spec with no mocks, it could be tagged as :slow to run separately once in a while.