add availability endpoint
tried to use ActiveModel::Attributes API for contract validation.
It has type casting and it will try to type cast string into datetime
for datetime params. If it fails passes nil down the line to the
presence validations. That's missleading because the param is not nil
but rather could not be casted into datetime. Type cast errors would be
appreciated instead.

==============================

Other than that the errors are all default, no locale was used.
Sorry don't have time to modify

==============================

I'll address controller specs later